rename email thread to email list card
